# CPP: Model taint through std::string and std::stringstream