--- a/cpp/ql/src/semmle/code/cpp/models/implementations/Strings.qll
+++ b/cpp/ql/src/semmle/code/cpp/models/implementations/Strings.qll
@@ -30,3 +30,39 @@ class StringCStr extends DataFlowFunction {
     output.isOutReturnValue()
   }
 }
+
+/**
+ * The standard function `operator<<`, for example on `std::stringstream`.
+ */
+class InsertionOperator extends TaintFunction {
+  InsertionOperator() {
+    this.hasQualifiedName("std", "operator<<")
+  }
+
+  override predicate hasTaintFlow(FunctionInput input, FunctionOutput output) {
+    (
+      // flow from second argument to first
+      input.isInParameterPointer(1) and
+      output.isOutParameterPointer(0)
+    ) or (
+      // flow from first argument to return value
+      input.isInParameterPointer(0) and
+      output.isOutReturnValue()
+    )
+  }
+}
+
+/**
+ * The standard function `std::stringstream.str`.
+ */
+class StringStreamStr extends TaintFunction {
+  StringStreamStr() {
+    this.hasQualifiedName("std", "basic_stringstream", "str")
+  }
+
+  override predicate hasTaintFlow(FunctionInput input, FunctionOutput output) {
+    // flow from object itself (qualifier) to return value
+    input.isInQualifier() and
+    output.isOutReturnValue()
+  }
+}