--- a/cpp/ql/test/library-tests/dataflow/taint-tests/models.expected
+++ b/cpp/ql/test/library-tests/dataflow/taint-tests/models.expected
@@ -1,0 +1,38 @@
+| stl.cpp:64:16:64:21 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 0 | OutReturnValue |
+| stl.cpp:64:16:64:21 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 1 | OutReturnValue |
+| stl.cpp:64:16:64:21 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 2 | OutReturnValue |
+| stl.cpp:65:16:65:24 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 0 | OutReturnValue |
+| stl.cpp:65:16:65:24 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 1 | OutReturnValue |
+| stl.cpp:65:16:65:24 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 2 | OutReturnValue |
+| stl.cpp:70:9:70:13 | call to c_str | stl.cpp:24:16:24:20 | c_str | dataflow | InQualifier | OutReturnValue |
+| stl.cpp:71:9:71:13 | call to c_str | stl.cpp:24:16:24:20 | c_str | dataflow | InQualifier | OutReturnValue |
+| stl.cpp:77:16:77:24 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 0 | OutReturnValue |
+| stl.cpp:77:16:77:24 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 1 | OutReturnValue |
+| stl.cpp:77:16:77:24 | call to basic_string | stl.cpp:22:3:22:14 | basic_string | dataflow | InParameter 2 | OutReturnValue |
+| stl.cpp:79:6:79:6 | call to operator<< | stl.cpp:42:67:42:76 | operator<< | taint | InParameterPointer 0 | OutReturnValue |
+| stl.cpp:79:6:79:6 | call to operator<< | stl.cpp:42:67:42:76 | operator<< | taint | InParameterPointer 1 | OutParameterPointer 0 |
+| stl.cpp:80:6:80:6 | call to operator<< | stl.cpp:42:67:42:76 | operator<< | taint | InParameterPointer 0 | OutReturnValue |
+| stl.cpp:80:6:80:6 | call to operator<< | stl.cpp:42:67:42:76 | operator<< | taint | InParameterPointer 1 | OutParameterPointer 0 |
+| stl.cpp:81:6:81:6 | call to operator<< | stl.cpp:42:67:42:76 | operator<< | taint | InParameterPointer 0 | OutReturnValue |
+| stl.cpp:81:6:81:6 | call to operator<< | stl.cpp:42:67:42:76 | operator<< | taint | InParameterPointer 1 | OutParameterPointer 0 |
+| stl.cpp:81:15:81:15 | call to operator<< | stl.cpp:42:67:42:76 | operator<< | taint | InParameterPointer 0 | OutReturnValue |
+| stl.cpp:81:15:81:15 | call to operator<< | stl.cpp:42:67:42:76 | operator<< | taint | InParameterPointer 1 | OutParameterPointer 0 |
+| stl.cpp:82:6:82:6 | call to operator<< | stl.cpp:43:85:43:94 | operator<< | taint | InParameterPointer 0 | OutReturnValue |
+| stl.cpp:82:6:82:6 | call to operator<< | stl.cpp:43:85:43:94 | operator<< | taint | InParameterPointer 1 | OutParameterPointer 0 |
+| stl.cpp:88:11:88:13 | call to str | stl.cpp:55:42:55:44 | str | taint | InQualifier | OutReturnValue |
+| stl.cpp:89:11:89:13 | call to str | stl.cpp:55:42:55:44 | str | taint | InQualifier | OutReturnValue |
+| stl.cpp:90:11:90:13 | call to str | stl.cpp:55:42:55:44 | str | taint | InQualifier | OutReturnValue |
+| stl.cpp:91:11:91:13 | call to str | stl.cpp:55:42:55:44 | str | taint | InQualifier | OutReturnValue |
+| taint.cpp:170:3:170:8 | call to strcpy | taint.cpp:156:7:156:12 | strcpy | dataflow | InParameter 0 | OutReturnValue |
+| taint.cpp:170:3:170:8 | call to strcpy | taint.cpp:156:7:156:12 | strcpy | dataflow | InParameterPointer 1 | OutParameterPointer 0 |
+| taint.cpp:170:3:170:8 | call to strcpy | taint.cpp:156:7:156:12 | strcpy | dataflow | InParameterPointer 1 | OutReturnPointer |
+| taint.cpp:172:3:172:8 | call to strcat | taint.cpp:157:7:157:12 | strcat | dataflow | InParameter 0 | OutReturnValue |
+| taint.cpp:172:3:172:8 | call to strcat | taint.cpp:157:7:157:12 | strcat | taint | InParameter 1 | OutParameterPointer 0 |
+| taint.cpp:172:3:172:8 | call to strcat | taint.cpp:157:7:157:12 | strcat | taint | InParameterPointer 0 | OutParameterPointer 0 |
+| taint.cpp:194:2:194:7 | call to memcpy | taint.cpp:190:7:190:12 | memcpy | dataflow | InParameter 0 | OutReturnValue |
+| taint.cpp:194:2:194:7 | call to memcpy | taint.cpp:190:7:190:12 | memcpy | dataflow | InParameterPointer 1 | OutParameterPointer 0 |
+| taint.cpp:194:2:194:7 | call to memcpy | taint.cpp:190:7:190:12 | memcpy | dataflow | InParameterPointer 1 | OutReturnPointer |
+| taint.cpp:194:2:194:7 | call to memcpy | taint.cpp:190:7:190:12 | memcpy | taint | InParameter 2 | OutParameterPointer 0 |
+| taint.cpp:194:2:194:7 | call to memcpy | taint.cpp:190:7:190:12 | memcpy | taint | InParameter 2 | OutReturnPointer |
+| taint.cpp:213:2:213:10 | call to swap | taint.cpp:201:35:201:38 | swap | dataflow | InParameterPointer 0 | OutParameterPointer 1 |
+| taint.cpp:213:2:213:10 | call to swap | taint.cpp:201:35:201:38 | swap | dataflow | InParameterPointer 1 | OutParameterPointer 0 |