--- a/cpp/ql/test/library-tests/dataflow/taint-tests/models.ql
+++ b/cpp/ql/test/library-tests/dataflow/taint-tests/models.ql
@@ -1,0 +1,19 @@
+import cpp
+import semmle.code.cpp.dataflow.TaintTracking
+import semmle.code.cpp.models.interfaces.DataFlow
+import semmle.code.cpp.models.interfaces.Taint
+
+from FunctionCall fc, Function f, string str, FunctionInput inModel, FunctionOutput outModel
+where
+	fc.getTarget() = f and
+	(
+		(
+			f.(TaintFunction).hasTaintFlow(inModel, outModel) and
+			str = "taint"
+		) or (
+			f.(DataFlowFunction).hasDataFlow(inModel, outModel) and
+			str = "dataflow"
+		)
+	)
+select
+	fc, f, str, inModel, outModel