Model taint on some very basic `std::string` and `std::stringstream` operations.  Enhance the AST taint library as required (@jbj please review changes to `TaintTracking.qll` and `DataFlowUtil.qll` when you get back; there was a little bit of experiment-until-it-seems-to-work-right involved in some of these changes).

For https://jira.semmle.com/browse/CPP-395 (though further work may be needed to resolve that issue).

In the long run I hope these models will benefit IR taint as well.