@geoffw0 2019-07-25
Performance seems fine (tested `NonConstantFormat.ql` on a snapshot of `MongoDB` (205s -> 227s) and Qt (493 -> 495s); no new results).