@geoffw0 2019-08-16
Fixed merge issues (tests and change note).