@geoffw0 2019-08-16
I haven't addressed all comments yet though.