@geoffw0 2020-04-02
Closed in favour of https://github.com/Semmle/ql/pull/3195 (the `operator<<` stuff will have to be repaired later).